pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omni-support-env"
version = "1.0.0"
description = "Multi-step customer operations RL environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "pydantic>=2.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "httpx>=0.27.0"]

[tool.hatch.build.targets.wheel]
packages = ["omni_support_env"]